requirements.txt
APScheduler==3.6.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
decorator==5.1.1
idna==3.10
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==13.0
pytz==2024.2
requests==2.32.3
six==1.16.0
telegram==0.0.1
tornado==6.4.1
tzlocal==5.2
urllib3==2.2.3
